Derive targetpath more reliably
The initial approach was to split the path by either \\ or / and locate
the first 'config' value in the list. But if the paths end up with both
slashes the script would break.

The second, quick and dirty approach was to simply split by 'config'
itself. But if the path contains config somewhere before the proper
config folder, it'll fail as well.

As such, I'm now splitting by the 'configtmp' path.. which should be
much more reliable.

I'll need to normalize the paths at some point and ensure the slashes
are all the same.. but I'll leave that for another time.