Change textbox logging method
The log method has been removed from view. Instead, LogBox now passes
two functions: log and get.

log: essentially does what view.log did, but also stores the messages in
a dict in the LogBox module. This allows for the old logs to be restored
when recreating a LogBox for a particular game.

get: returns the textbox widget itself, which is needed to add the
widget to the grid and position it.