Remove uac in bundle; implement elevate
The app will now request UAC upon launch if one or more of the install
paths require admin privileges. If the user selects a path that will
need UAC, the app spawns a modal telling the user to restart.

I had originally tried to get elevate to prompt the user and spawn a new
instance of the launcher when they selected a path. However, the old
instance seemed to freeze out rather than simply closing. Not sure why..
hence the info modal.