Delete any widget that is not cover_frame
An issue was occurring whenever you'd switch modpacks with the settings
tab open. The mouse would hover one of the setting inputs and cause an
infinitely persisting ToolTip for it -- even after the element has been
destroyed.

This will ensure that anything that is not the cover frame will be
cleaned up as well.