Copy local files instead of moving them
Some users may want to recover their old instance. Moving them, instead
of copying them, will make that process much harder for them.. as such,
it's best to just copy them (despite the space increase).

Ideally, the user won't select an instance that already has a modpack
installed and the space increase won't be an issue in the first place.