Working Fallout76Updater
There is an issue with `install_update`; if the client already has some
of the mods it will fail due to `shutil.copytree` saying permission
denied. It seems to have a particular issue with ReShade's `.undef`
files.

I may just add an `initial_install` that checks if ReShade exists
already and delete it. But it may be better to just create a custom copy
function for `shutil.copytree` and handle exceptions on a per-file
basis.