Begin work on automatic path detection
Currently has support for finding window special folders (i.e. My
Documents), Steam games, and games that can be found in the Start
Menu.

The Start Menu code isn't fully functional yet. It works with some
but not all apps. The AppID can be an identifier or the direct path
to an exe; it works for exes but not identifiers.

Needs some work still.