Clean up path finder code
The code is much more modular now and works based on AppID conventions.
This simplifies the code as you only need to use `find_appid_path` to
find steam game paths, microsoft app paths, and other start menu paths.

In addition, the steam game path finder is now more intelligent and does
not require a "directory" to be passed to it. It will look for the
game's `appmanifest` file to determine the directory name.